+++ examples/grouped_failure/ballast/module_05.py
@@ -1,0 +1,3 @@
+"""Independent ballast module 05."""
+
+VALUE = 5